+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 0.4.0
 
-- Search deep inside node modules (limited to those that are specified in dependencies).
+- Search deep inside node modules (limited to those that are specified in dependencies). **Disabled** per default.
 - Improve the info given about each file.
 - FIX: if trying to require the current file, show a warning instead of creating a broken reference.